fixed log

+++ application/ylp-activities/model/local/activity_yhb_member.js
@@ -1,6 +1,6 @@
 module.exports = {
     attributes: {
         aid:{type:'string',index:true,required:true}
-        ,redpack:{type:'integer',defaultsTo:0}
+        ,redpack:{type:'float',defaultsTo:0}
     }
 }
